cli: in batch mode, print empty line if get fails
Currently get commands return nothing to stdout on error in batch mode.
In practice, this makes them hard to use in scripts. If the caller
sends 4 gets and we only return 3 lines, the caller can not easily
determine which one did not work.

This changes the get command behavior in batch mode to always print
a new line, even on error. This allows the caller to determine which
get did not work. e.x. if the 3rd line is blank then the 3rd get
did not work.

Signed-off-by: Matthew Cather <mattbob4@gmail.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>